requirements.txt
python == 3.6

altgraph==0.17
fpdf==1.7.2
imutils==0.5.3
numpy==1.18.3
opencv-contrib-python==4.2.0.34
pillow==7.1.2
pyinstaller==3.6
pytesseract==0.3.4
scipy==1.4.1
PyQt5==5.14.2 
sqlparse==0.3.1


